rename: component 하위 디렉토리 생성
