+++ app/components/bookshelf/book-thumbnail.tsx
